% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{add_missing_nodes}
\alias{add_missing_nodes}
\title{Add nodes for flows without a start or end node}
\usage{
add_missing_nodes(dat, flow_direction = "var_fout")
}
\arguments{
\item{dat}{Tibble.}

\item{flow_direction}{String. "var_fout" to look for missing end nodes,
"var_fin" to look for missing start nodes.}
}
\value{
Input tibble with missing nodes appended.
}
\description{
Search flows labelled by a flow direction for missing counternode and
add row data as required. The magnitude of the flow will be the sum of
all incoming flows
}
\keyword{internal}
