% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{add_missing_nodes_subfunction}
\alias{add_missing_nodes_subfunction}
\title{Return tibble with missing nodes appended as required}
\usage{
add_missing_nodes_subfunction(dat, direction, commodity)
}
\arguments{
\item{dat}{Tibble from nested data by commodity.}

\item{direction}{String "var_fout" or "var_fin".}

\item{commodity}{String. Grouping variable in add_missing_nodes.}
}
\value{
Input tibble with missing nodes inserted.
}
\description{
Return tibble with missing nodes appended as required
}
\keyword{internal}
