% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{assign_node_num}
\alias{assign_node_num}
\title{Append node numbers to dat}
\usage{
assign_node_num(dat, nodes, col_to_assign_num)
}
\arguments{
\item{dat}{Tibble - long.}

\item{nodes}{Tibble output of make_nodes.}

\item{col_to_assign_num}{Column in dat to use for
assigning node number}
}
\value{
Input tibble with node_num appended
}
\description{
Append node numbers to dat
}
\keyword{internal}
