% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{check_in_path}
\alias{check_in_path}
\title{Check whether a regular expression is present in nodes (processes) in a list of paths}
\usage{
check_in_path(node_regex, path)
}
\arguments{
\item{node_regex}{A string regular expression to search for in process nodes.}

\item{path}{A list of paths}
}
\value{
logical list
}
\description{
Check whether a regular expression is present in nodes (processes) in a list of paths
}
\examples{
 dg <- demos_001_sector \%>\%
           filter(period == 2006) \%>\%
           make_graph_from_veda_df(node_labels = process,
                         edge_labels = commodity
                           )
 all_mincoa1_paths <- all_simple_paths(g, from = "mincoa1")
 check_in_paths("(exp)", all_min_coa1_paths)
}
