% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_veda_dat.R
\name{define_sector_from_list}
\alias{define_sector_from_list}
\title{define sectors in imported veda data from list}
\usage{
define_sector_from_list(
  dat,
  join_variable_name,
  sector_dat,
  sector_info_column,
  sector_dat_join_variable_col
)
}
\arguments{
\item{dat}{A tibble of veda data from import_vd or prep_data}

\item{join_variable_name}{A string column name in dat for joining to sector_dat
export}

\item{sector_dat}{A tibble containing the sector information for variables}

\item{sector_info_col}{Column in sector_dat containing sector information}

\item{sector_dat_variable_col}{Column in sector_dat for joining to dat}
}
\description{
Append sector information from a tibble to tibble output from
\code{import_vd()} or \code{prep_data()}. The function joins the sector
 information based on a joining variable
}
\examples{
test_dat <- tibble::tibble(attribute = "var_fout",
             commodity = "adistelc00",
             process = "adistelc00" ,
             period = NA,
             vintage = NA,
             timeslice = "annual",
             region = "reg1",
             userconstraint = "-",
             pv = 10,
             commodity_description = NA,
             process_description = NA,
             usercostraint_description = NA,
             commodity_set = NA,
             process_set = NA,
             usercostraint_set = NA)

sector_dat <- tibble::tibble(major_sector` = "agr",
                          process = "adistlec00")

t <- define_sector_from_list(test_dat, "process", sector_dat,
                            major_sector, process)

export
}
