% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_veda_dat.R
\name{define_sector_from_string}
\alias{define_sector_from_string}
\title{Define sectors based on string matches}
\usage{
define_sector_from_string(dat)
}
\arguments{
\item{dat}{A tibble of veda data from import_vd or prep_data}
}
\description{
Use specified regular expressions for sector definitions based on naming
conventions used for Scottish Times Model
}
