% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demos_001}
\alias{demos_001}
\title{Results from demos_001 run created by prep_data()}
\format{
Tibble with 117 rows and 15 columns:
\describe{
    \item{attribute}{attribute name as lower case string}
    \item{commodity}{commodity name as lower case string}
    \item{process}{process name as lower case string}
    \item{period}{model period as numeric}
    \item{region}{region name as lower case string}
    \item{vintage}{model vintage as numeric}
    \item{timeslice}{timeslice as string. Missing data converted to
    annual}
    \item{usercontraint}{userconstraint name as lower case string}
    \item{pv}{numeric value for the datum}
    \item{commodity_description}{Imported commodity description as
    lower case string}
     \item{process_description}{Imported process description as
    lower case string}
     \item{userconstraint_description}{Imported userconstraint description as
    lower case string}
     \item{commodity_set}{Imported commodity sets as
    lower list object}
    \item{process_set}{Imported process sets as
    lower list object}
    \item{userconstriant_set}{Imported userconstraint sets as
    lower list object}
}


}
\usage{
demos_001

demos_001
}
\description{
Results data from the DemoS_001 model in veda 2 using demos_001 run.
Dataset containing all attributes, commodities and processes#'

Results data from the DemoS_001 model in veda 2 using demos_001 run
with sectors defined from list. Dataset containing all
attributes, commodities and processes and sector appended#'
}
\keyword{datasets}
