% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demos_001_results}
\alias{demos_001_results}
\title{Exported results tables of DemoS_001 runs in Veda2}
\format{
A list of 8 elements:
\describe{
    \item{All costs.csv}{Tibble (13 x 12) cost_* attributes for coal processes}
    \item{Annual technology costs.csv}{Tibble (4 x 11) Cost_FOM, Cost_Inv
    for Process == "DTPSCOA"}
    \item{Demands.csv}{Tibble (2 x 13) VAR_FOut demand commodities TPSCOA}
    \item{Fuel Supply.csv}{Tibble (6 x 13) VAR_FOut for IMPCOA and MINCOA
    processes}
    \item{Prices_All.csv}{Tibble (4 x 10) EQ_CombalM for DTPCOA and COA}
    \item{Prices_Energy.csv}{Tibble (2 x 11) EQ_CombalM for COA}
    \item{Process Marginals.csv}{Tibble (11 x 11) VAR_ActM for processes}
    \item{_SysCost.csv}{Tibble (1 x 7) Reg_obj}

}
}
\usage{
demos_001_results
}
\description{
A list object in which each list corresponds to one of the csvs
created by the batch export function in Veda2
}
\keyword{datasets}
