% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demos_001_vdt}
\alias{demos_001_vdt}
\title{Tibble produced by prep_vdt_dat("demos_001")}
\format{
Tibble with 11 rows and 6 columns:
\describe{
    \item{region}{region name as lower case string}
    \item{process}{process name as lower case string}
    \item{commodity}{commodity name as lower case string}
    \item{direction}{"in" or "out"}
     \item{process_description}{Imported process description as
    lower case string}
    \item{commodity_description}{Imported commodity description as
    lower case string}
}
}
\usage{
demos_001_vdt
}
\description{
Tibble of DemoS_001 model vdt data
}
\keyword{datasets}
