% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demos_007_vdt}
\alias{demos_007_vdt}
\title{Tibble produced by prep_vdt_dat("demos_007")}
\format{
Tibble with 375 rows and 6 columns:
\describe{
    \item{region}{region name as lower case string}
    \item{process}{process name as lower case string}
    \item{commodity}{commodity name as lower case string}
    \item{direction}{"in" or "out"}
     \item{process_description}{Imported process description as
    lower case string}
    \item{commodity_description}{Imported commodity description as
    lower case string}
}
}
\usage{
demos_007_vdt
}
\description{
Tibble of DemoS_007 model vdt data
}
\keyword{datasets}
