% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_res.R
\name{make_edges}
\alias{make_edges}
\title{Create tibble of edge data from long tibble}
\usage{
make_edges(dat, node_col, flow_col)
}
\arguments{
\item{dat}{Tibble - long.}

\item{node_col}{The column to use for node information.
networkD3 requires node_num.}

\item{flow_col}{Column that specifies ID of unique flows (commodity).}
}
\value{
Tibble with source-target pairs for each flow.
}
\description{
Create tibble of edge data from long tibble
}
\keyword{internal}
