% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{make_nodes}
\alias{make_nodes}
\title{Extract all nodes from dat}
\usage{
make_nodes(dat, node_column)
}
\arguments{
\item{dat}{Tibble - long.}

\item{node_column}{The column in dat for getting the node identities.}
}
\value{
Tibble with node_column and node_num.
}
\description{
Extract nodes from dat in long format and assign
zero indexed node_num
}
\keyword{internal}
