% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_res.R
\name{make_res}
\alias{make_res}
\title{Create a RES sankey diagram with data for a specified sector and period}
\usage{
make_res(
  dat,
  period_select = NULL,
  sector_select = NULL,
  region_select = NULL,
  node_labels = process_description,
  edge_labels = commodity_description,
  font_size
)
}
\arguments{
\item{dat}{Tibble output from prep_data() \%>\% define_sector_from_*().}

\item{period_select}{Numeric. Period for plotting RES.}

\item{sector_select}{String. Sector for plotting RES.}

\item{region_select}{String. Region for plotting RES}

\item{node_labels}{Column in dat for labelling nodes.}

\item{edge_labels}{Column in dat for labelling edges.}

\item{font_size}{Numeric. Font size for RES labels.}
}
\value{
NetworkD3 Sankey object
}
\description{
Use the full dataset from prep_data %>% define_sector_*() to create a sankey
diagram linking processes with commodities shown as flows. Flow magnitude
information is not included. The RES is shown for an individual
specified period and sector. The nodes are labelled with process_description.
This will be changed to an option
}
\examples{
 data(demos_001_sector)
 demos_001_sector \%>\%
    make_res(period_select = 2005, sector_select = "coal",
             node_labels = process_description,
             edge_labels = commodity_description,
             font_size = 11)
}
