% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{make_sankey}
\alias{make_sankey}
\title{Make NetworkD3 sankey object}
\usage{
make_sankey(
  nodes,
  edges,
  source,
  target,
  value,
  node_label = process_description,
  edge_label = NULL,
  sankey_width = NULL,
  sankey_height = NULL,
  font_size = 12
)
}
\arguments{
\item{nodes}{Nodes tibble zero indexed.}

\item{edges}{Edge data with source and target zero-indexed.}

\item{source, }{target, value Column names in edges tibble.}

\item{edge_label, }{node_label Enquoted columns in edges tibble. Used for flow
tooltip}

\item{sankey_width}{Width (in pixels) of sankey.}

\item{sankey_height}{Height (in pixels) of sankey.}

\item{font_size}{Numeric.}
}
\value{
NetworkD3
}
\description{
Make NetworkD3 Sankey object from node and edge data with tooltips
}
\keyword{internal}
