% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_veda_dat.R
\name{prep_data}
\alias{prep_data}
\title{Import output from veda runs into R.}
\usage{
prep_data(filename_base, vignette = F)
}
\arguments{
\item{filename_base}{a character string of the veda files without the .vd* suffix}

\item{vignette}{logical. Set to TRUE if running from within the vignette or development environmnet.}
}
\description{
Imports and combines vd, vde and vds files to create a
 single tibble with standard formatting
}
\examples{
  path_to_data <- paste0(find.package("vedar"),
                         "/extdata/")
  filename_base <- paste0(path_to_data, "DemoS_001")

  dat <- prep_data(filename_base)
}
