% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{syscost}
\alias{syscost}
\title{Extract the undiscounted system cost}
\usage{
syscost(dat)
}
\arguments{
\item{dat}{Tibble of veda output data from prep_data with or without
additional variables}
}
\value{
Tibble (n(regions) x 2): region, syscost
}
\description{
Extract undiscounted system cost as the reg_obj variable for each region
}
