% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{total_target_var_fin_by_source_function}
\alias{total_target_var_fin_by_source_function}
\title{Compute the sum of the var_fin over targets for each source by commodity}
\usage{
total_target_var_fin_by_source_function(dat, source_val, commodity_val)
}
\arguments{
\item{dat}{Tibble of source-target edge data}

\item{source_val}{String value of source}

\item{commodity_val}{String value of commodity}
}
\value{
numeric total of var_fin over target by source, commodity
}
\description{
Compute the sum of the var_fin over targets for each source by commodity
}
\keyword{internal}
