% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_veda_dat.R
\name{vd_structure_match_expected}
\alias{vd_structure_match_expected}
\title{Check structure of vd data}
\usage{
vd_structure_match_expected(filename, filetype)
}
\arguments{
\item{filename}{String. vd filename.}

\item{filetype}{String "vd_file", "vde_file", "vds_file"}
}
\value{
TRUE if structure matches expected  structure
}
\description{
Check the input data structure is as expected
}
