% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{make_res}
\alias{make_res}
\title{Use the full dataset from prep_data %>% define_sector_*() to create a sankey
diagram linking processes with commodities shown as flows. Flow magnitude
information is not included. The RES is shown for an individual
specified period and sector. The nodes are labelled with process_description.
This will be changed to an option}
\usage{
make_res(
  dat,
  node_labels = process_description,
  edge_labels = commodity_description,
  sankey_width = NULL,
  sankey_height = NULL,
  font_size = 10,
  use_weights = T,
  input_data_type = "vd"
)
}
\arguments{
\item{dat}{Tibble output from prep_data() \%>\% define_sector_from_*().}

\item{node_labels}{Column in dat for labelling nodes.}

\item{edge_labels}{Column in dat for labelling edges.}

\item{sankey_width}{Width (in pixels) of sankey.}

\item{sankey_height}{Height (in pixels) of sankey.}

\item{font_size}{Numeric. Font size for RES labels.}

\item{use_weights}{Logical. Whether to represent edge weights in Sankey that contains data for a single period. For data containing more than 1 period, magnitudes are unitary.}

\item{input_data_type}{Character. "vd" or "vdt"}
}
\value{
NetworkD3 Sankey object
}
\description{
Use the full dataset from prep_data %>% define_sector_*() to create a sankey
diagram linking processes with commodities shown as flows. Flow magnitude
information is not included. The RES is shown for an individual
specified period and sector. The nodes are labelled with process_description.
This will be changed to an option
}
\examples{
 data(demos_001_sector)
 demos_001_sector \%>\%
    make_res(period_select = 2005, sector_select = "coal",
             node_labels = process_description,
             edge_labels = commodity_description,
             font_size = 11)

 data(demos_001_vdt)
 make_res(demos_001_vdt \%>\%
             filter(region == "reg1"),
         input_data_type = "vdt")

}
